Build sklearnserver
Execution (finally) gives the same output as the upstream image,
which is admittedly a Python exception because there's no
JOBLIB_FILE from some `seldon_core...model_uri`, but it's close
to a working model if not completely there.

Primarily, Conda inside craft is an adventure for a lot of reasons,
and while this is the first Rockcraft YAML which successfully jumps
through all the right hoops, the patterns here (principally the
subshell, invoking `seldon-core-microservice ..`, another subshell
during pebble exec) will almost certainly be needed in other Seldon
images, and/or other Conda images.