chore: update rockcraft project, add sanity test, and tox.ini file
This commit updates the rockcraft project to match the new requirements from the
latest rockcraft version. It also refactors the security requirement to match the latest
format in other rock repositories.
Bash is now removed from the image as it is not required and not recommended in production
environments, and the build dependency on apt is also removed as it was not used at all.
Adding sanity tests in this rock's repository and its corresponding tox file.

Fixes #71